Move group related entities to group package
