+++ ippon/group/serializers.py
@@ -1,0 +1,13 @@
+from rest_framework import serializers
+
+import ippon.models
+
+
+class GroupSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = ippon.models.group.Group
+        fields = (
+            'id',
+            'name',
+            'group_phase'
+        )
