+++ ippon/group_phase/views.py
@@ -5,7 +5,7 @@ from rest_framework.response import Response
 
 from ippon.authorization_views import has_tournament_authorization
 from ippon.models import GroupPhase, Group
-from ippon.serializers import GroupSerializer
+from ippon.group.serializers import GroupSerializer
 from ippon.group_phase.serializers import GroupPhaseSerializer
 from ippon.tournament import permissions as tp
