+++ ippon/serializers.py
@@ -15,16 +15,6 @@ class MinimalUserSerializer(serializers.ModelSerializer):
         )
 
 
-class GroupSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = ippon.models.group.Group
-        fields = (
-            'id',
-            'name',
-            'group_phase'
-        )
-
-
 class GroupFightSerializer(serializers.ModelSerializer):
     class Meta:
         model = ippon.models.group_fight.GroupFight
